+++ main-app11/main-app/preload.js
@@ -105,6 +105,7 @@ contextBridge.exposeInMainWorld('electronAPI', {
   videoMakeProxy: (opts) => ipcRenderer.invoke('video-make-proxy', opts),
   videoCancelProxy: () => ipcRenderer.invoke('video-cancel-proxy'),
   videoTimelineAssets: (opts) => ipcRenderer.invoke('video-timeline-assets', opts),
+  videoAudioStems: (opts) => ipcRenderer.invoke('video-audio-stems', opts),
   onVideoExportProgress: (callback) => ipcRenderer.on('video-export-progress', (_event, data) => callback(data)),
   onVideoProxyProgress: (callback) => ipcRenderer.on('video-proxy-progress', (_event, data) => callback(data)),